% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwich2stage.R
\name{sandwich2stage}
\alias{sandwich2stage}
\title{Implement sandwich variance estimation in 2-stage regression.}
\usage{
sandwich2stage(
  stage1.model,
  stage2.model,
  xstar = "xstar",
  xhat = "xhat",
  Stage1ID = "ID",
  Stage2ID = "ID"
)
}
\arguments{
\item{stage1.model}{the fitted stage 1 (calibration) model object fit using \code{svyglm()}}

\item{stage2.model}{the fitted stage 2 (outcome) model object fit using \code{svyglm()} or \code{svycoxph()}}

\item{xstar}{the name of the error-prone variable in the main study data}

\item{xhat}{the name of the estimated exposure variable from the stage 1 model}

\item{Stage1ID}{the name of the subject-level ID variable in the data used to fit the stage 1 model}

\item{Stage2ID}{the name of the subject-level ID variable in the data used to fit the stage 2 model}
}
\value{
returns a list containing the stage 1 and stage 2 model calls and formulas, the vector of estimated coefficients from the stage 1 and stage 2 models, the estimated sandwich variance matrix, and the matrix of influence functions.
}
\description{
This function provides an estimate of the sandwich variance from the 2-stage regression approach of regression calibration, which is obtained by stacking the stage 1 and stage 2 model estimating equations.
}
\details{
This function can be used for a linear stage 1 (calibration) model and when the stage 2 model is a linear, GLM, or Cox regression model. This function requires the following arguments: the fitted stage 1 and 2 model objects (fit using functions from the \code{survey} package, e.g. \code{svyglm()}), the names of the error-prone and estimated exposure variables in the data, and the names of the subject ID variables in the data sets used to fit the stage 1 and 2 models.
}
\examples{
#Begin by loading our package and then loading the data sandwichdata_SRS.
library(sandwich2stage)
data("sandwichdata_SRS")
#Next, we will create a survey design object (e.g. simple random sample).
sampdesign <- survey::svydesign(id=~1, data=sandwichdata_SRS)
#We may then fit the stage 1 and stage 2 models, saving the estimated nuisance parameters from the
#stage 1 model and using them to obtain an estimated of the unknown exposure (xhat), which will be
#used in the stage 2 model. We will then fit out stage 2 outcome model.
stage1.model<-survey::svyglm(xstarstar~xstar+z,design=sampdesign,family=gaussian(),subset=v==1)
alphas.stage1<-coef(stage1.model)
sampdesign <- update(sampdesign,xhat =predict(stage1.model,newdata=sampdesign$variables) )
stage2.model<-  survey::svyglm(y ~ xhat+z,design=sampdesign,family=binomial())
#Finally, we can obtain an estimate of the sandwich variance using our 2-step regression procedure.
sandwich.object<-sandwich2stage(stage1.model,stage2.model,
xstar="xstar",xhat="xhat",
Stage1ID="ID",Stage2ID="ID")
sandwichvar<-vcov(sandwich.object)
}
\references{
Binder, D. A. (1983). On the variances of asymptotically normal estimators from complex surveys. International Statistical Review/Revue Internationale de Statistique, pages 279–292.

Boos, D. and Stefanski, L. (2013). Essential Statistical Inference: Theory and Methods. Springer, New York, NY.

Lumley, T. (2011). Complex surveys: a guide to analysis using R, volume 565. John Wiley & Sons.

Lumley, T. and Scott, A. (2017). Fitting regression models to survey data. Statistical Science, pages 265–278.
}
